refactor: seed exec domain package
