+++ pkg/agentcompose/service.go
@@ -1,6 +1,7 @@
 package agentcompose
 
 import (
+	execdomain "agent-compose/internal/agentcompose/exec"
 	appconfig "agent-compose/pkg/config"
 	driverpkg "agent-compose/pkg/driver"
 	"context"
@@ -1047,64 +1048,18 @@ func parseAgentExecResult(agent string, result ExecResult) (AgentRunResult, erro
 }
 
 func agentTraceEvents(transcript string, createdAt time.Time) []SessionEvent {
-	lines := strings.Split(transcript, "\n")
-	events := make([]SessionEvent, 0)
-	for index := 0; index < len(lines); index++ {
-		line := strings.TrimSpace(lines[index])
-		eventType, name, ok := parseAgentTraceMarker(line)
-		if !ok {
-			continue
-		}
-		details, consumed := collectAgentTraceDetails(eventType, lines[index+1:])
-		index += consumed
-		message := name
-		if strings.TrimSpace(details) != "" {
-			if message == "" {
-				message = strings.TrimSpace(details)
-			} else {
-				message += "\n" + strings.TrimSpace(details)
-			}
-		}
+	domainEvents := execdomain.AgentTraceEvents(transcript)
+	events := make([]SessionEvent, 0, len(domainEvents))
+	for _, event := range domainEvents {
 		events = append(events, SessionEvent{
 			ID:        uuid.NewString(),
-			Type:      eventType,
-			Level:     "info",
-			Message:   message,
+			Type:      event.Type,
+			Level:     event.Level,
+			Message:   event.Message,
 			CreatedAt: createdAt,
 		})
 	}
 	return events
-}
-
-func collectAgentTraceDetails(eventType string, lines []string) (string, int) {
-	details := make([]string, 0, len(lines))
-	for offset, raw := range lines {
-		line := strings.TrimSpace(raw)
-		if _, _, marker := parseAgentTraceMarker(line); marker {
-			return strings.Join(details, "\n"), offset
-		}
-		if eventType != "agent.assistant" && line == "" {
-			return strings.Join(details, "\n"), offset + 1
-		}
-		details = append(details, raw)
-	}
-	return strings.Join(details, "\n"), len(lines)
-}
-
-func parseAgentTraceMarker(line string) (string, string, bool) {
-	if strings.HasPrefix(line, "[tool:") && strings.HasSuffix(line, "]") {
-		name := strings.TrimSpace(strings.TrimSuffix(strings.TrimPrefix(line, "[tool:"), "]"))
-		if name != "" {
-			return "agent.tool", name, true
-		}
-	}
-	if strings.HasPrefix(line, "[hook:") && strings.HasSuffix(line, "]") {
-		name := strings.TrimSpace(strings.TrimSuffix(strings.TrimPrefix(line, "[hook:"), "]"))
-		if name != "" {
-			return "agent.hook", name, true
-		}
-	}
-	return "", "", false
 }
 
 func validateLoaderCommandRequest(request LoaderCommandRequest) error {
@@ -1404,14 +1359,7 @@ func buildAgentExecSpec(config *appconfig.Config, session *Session, agent, model
 }
 
 func summarizeAgentResult(result AgentRunResult) string {
-	body := firstNonEmpty(result.FinalText, result.DisplayOutput, result.Transcript)
-	if strings.TrimSpace(body) == "" {
-		if result.Success {
-			return fmt.Sprintf("%s finished without output", result.Agent)
-		}
-		return fmt.Sprintf("%s failed without output", result.Agent)
-	}
-	return body
+	return execdomain.SummarizeAgentResult(result)
 }
 
 func toProtoSessionDetail(session *Session) *agentcomposev1.SessionDetail {
